Ermine.Builtin.Pattern

Description

Smart builders for convenient building of patterns.

Documentation

data Binder v a

Constructors

Binder

Fields

vars :: [v]

item :: a

note :: Binder v v -> Binder v v

If the annotation and item types correspond, we can push the item onto the annotations

noted :: v -> Binder v v

Injects a value into a binder, placing the value in the annotations as well.

type P t v = Binder v (Pattern t)

Smart pattern

varp :: v -> P (Annot k t) v

A pattern that binds a variable with (effectively) no type annotation.

sigp :: v -> t -> P t v

A pattern that binds a variable with a type annotation.

_p :: P t v

A wildcard pattern that ignores its argument

strictp :: P t v -> P t v

A strict (bang) pattern

lazyp :: P t v -> P t v

A lazy (irrefutable) pattern

asp :: v -> P t v -> P t v

An as (@) pattern.

conp :: Global -> [P t v] -> P t v

A pattern that matches a constructor expression, with a specified number of unboxed arguments.

litp :: Literal -> P t v

A pattern that matches a literal unboxed value

alt :: (Monad f, Eq v) => P t v -> Guarded (f v) -> Alt t f v

smart alt constructor