Ermine.Builtin.Type

Description

This module supplies a number of builtin types intrinsic to the Ermine language.

>>> infer int
*

>>> infer (list list)
*** Exception: kind mismatch

Builtin Types

class Builtin t where

This class allows for overloading of builtin types, so they can be used directly a nice DSL for specifying types borrowing application from Haskell for the application of types.

Methods

builtin :: Ord k => Kind k -> String -> t

builtin_ :: Builtin t => String -> t

Create a Builtin Type of Kind star.

Numerics

int :: Builtin t => t

>>> infer int
*

long :: Builtin t => t

>>> infer long
*

byte :: Builtin t => t

>>> infer byte
*

short :: Builtin t => t

>>> infer short
*

float :: Builtin t => t

>>> infer float
*

double :: Builtin t => t

>>> infer double
*

Booleans

bool :: Builtin t => t

>>> infer bool
*

Text

string :: Builtin t => t

>>> infer string
*

char :: Builtin t => t

>>> infer char
*

Containers

list :: Builtin t => t

>>> infer list
* -> *

>>> infer (list int)
*

>>> infer (tup [int, list int])
*

maybe_ :: Builtin t => t

>>> infer maybe_
* -> *

>>> infer (maybe_ int)
*

ee :: Builtin t => t

>>> infer ee
*

Type Equality

equality :: Builtin t => t

>>> infer equality
a -> a -> *

>>> infer (equality equality)
(a -> a -> *) -> *

Classes

functor :: Builtin t => t

>>> infer functor
(* -> *) -> Γ

>>> infer (equality functor)
((* -> *) -> Γ) -> *

lame :: Builtin t => t

This is a placeholder typeclass we're using until we get class declaration parsing implemented

>>> infer lame
* -> Γ

Annotations

anyType :: Annot t a

A type annotation that can be applied to anything.