Ermine.Console.Unicode

Description

Documentation

withUnicode :: MonadCatchIO m => m a -> m a

Run in a modified codepage where we can print UTF-8 values on Windows.

You should probably run the top level of your program in this.