Ermine.Diagnostic

Description

This module re-exports Rendering from trifecta along with our in-house diagnostic type.

This allows the non-parsing modules to not have to concern themselves with how we choose to represent source locations.

Rendering

data Rendering

class HasRendering t0 where

Methods

rendering :: Lens' t0 Rendering

renderingDelta :: Lens' t0 Delta

renderingLine :: Lens' t0 (Lines -> Lines)

renderingLineBytes :: Lens' t0 Int64

renderingLineLen :: Lens' t0 Int64

renderingOverlays :: Lens' t0 (Delta -> Lines -> Lines)

Diagnostics

data Diagnostic

Ermine diagnostic type

Constructors

Diagnostic !Rendering (Maybe Doc) [Doc] (Set String)

class AsDiagnostic t where

This provides the _Diagnostic prism that can be used when matching against SomeException or in a custom Error type.

Methods

_Diagnostic :: Prism' t Diagnostic

class HasDiagnostic t where

Methods

diagnostic :: Lens' t Diagnostic

diagnosticExpected :: Lens' t (Set String)

diagnosticMessage :: Lens' t (Maybe Doc)

diagnosticNotes :: Lens' t [Doc]

diagnosticRendering :: Lens' t Rendering

data Err

Constructors

Err (Maybe Doc) [Doc] (Set String)

class HasErr t0 where

Methods

err :: Lens' t0 Err

expected :: Lens' t0 (Set String)

footnotes :: Lens' t0 [Doc]

reason :: Lens' t0 (Maybe Doc)

die :: HasRendering r => r -> Text -> Diagnostic

Construct a diagnostic