Ermine.Digest

Description

MD5 Digests

data Digest

A 128-bit MD5 digest

>>> digest "hello"
Digest 10268287928962245903 5020952346831662890

Constructors

Digest !Word64 !Word64

Constructing Digests

class MD5 t where

A class for things that can be converted directly to an MD5 digest.

Methods

digest :: t -> Digest

compute the digest. This will default to computing the digest using show if not specified

class MD5List t where

A helper class provided to aid in MD5 calculation

Methods

digestList :: [t] -> Digest