Ermine.Mangled

Description

Classes and tools for traversing and binding.

Bound by another type

class Monad m => BoundBy tm m | tm -> m where

Generalizes Bound to permit binding by another type without taking it as a parameter.

Methods

boundBy :: (a -> m b) -> tm a -> tm b

Mangling

class (Traversable t, Applicative t, Monad t) => Mangled t where

Perform simultaneous traversal and binding.

Methods

mangled :: Applicative f => (a -> f (t b)) -> t a -> f (t b)

traverseDefault :: (Mangled t, Applicative f) => (a -> f b) -> t a -> f (t b)

A default definition of traverse defined in terms of mangled

bindDefault :: Mangled t => t a -> (a -> t b) -> t b

A default definition of (>>=) defined in terms of mangled.

Mangling by another type

class (Traversable tm, Mangled m, BoundBy tm m) => MangledBy tm m | tm -> m where

Generalizes Mangled to permit binding by another type without taking it as a parameter.

Methods

mangledBy :: Applicative f => (a -> f (m b)) -> tm a -> f (tm b)