Ermine.Match

Description

Documentation

data Claused c a

Constructors

Localized [Scope PatternPath (Scope Word32 c) a] (Guarded (Scope PatternPath (Scope Word32 c) a))

Raw (Guarded (Scope PatternPath c a))

data PatternMatrix t c a

Pattern matrices for compilation. The matrix is represented as a list of columns. There is also an extra column representing the guards.

Constructors

PatternMatrix

Fields

_cols :: [[Pattern t]]

_bodies :: [Claused c a]

class HasPatternMatrix t t c a | t -> t c a where

Minimal complete definition

patternMatrix

Methods

patternMatrix :: Lens' t (PatternMatrix t c a)

bodies :: Lens' t [Claused c a]

cols :: Lens' t [[Pattern t]]

newtype MatchEnv

The environment necessary to perform pattern compilation. We need two pieces of information:

1) The signatures (full list of constructors) associated with each particular constructor

2) A mapping from constructors to their Core integer tag

This is accomplished via a map of maps. The outer map should take each global to its associated signature, and signatures are represented as maps from globals to integer tags.

Constructors

MatchEnv

Fields

signatures :: HashMap Global (HashMap Global Word8)

dummyMatchEnv :: MatchEnv

class (Applicative m, Monad m) => MonadMatch m where

Monads that allow us to perform pattern compilation, by providing a MatchEnv.

Methods

askMatch :: m MatchEnv

isSignature :: MonadMatch m => Set PatternHead -> m Bool

Determines whether a set of pattern heads constitutes a signature. This is handled specially for tuples and literals, and relies on the monad for data type constructors.

constructorTag :: MonadMatch m => PatternHead -> m Word8

Looks up the constructor tag for a pattern head. For tuples this is always 0, but constructors must consult the compilation environment.

data Matching c a

Additional information needed for pattern compilation that does not really belong in the pattern matrix.

The pathMap stores information for resolving previous levels of the pattern to appropriate cores for the current context.

colCores contains core terms that refer to each of the columns of the matrix in the current context.

colPaths contains the paths into the original pattern that correspond to each of the current columns.

Constructors

Matching

Fields

_pathMap :: HashMap PatternPath (c a)

_colCores :: [c a]

_colPaths :: [PatternPaths]

class HasMatching t c a | t -> c a where

Minimal complete definition

matching

Methods

matching :: Lens' t (Matching c a)

colCores :: Lens' t [c a]

colPaths :: Lens' t [PatternPaths]

pathMap :: Lens' t (HashMap PatternPath (c a))

defaultOn :: Applicative c => Int -> PatternMatrix t c a -> PatternMatrix t c (Var () (c a))

Computes the matrix that should be used recursively when defaulting on the specified column.

splitOn :: Applicative c => Int -> PatternHead -> PatternMatrix t c a -> PatternMatrix t c (Var Word8 (c a))

Computes the matrix that should be used recursively when defaulting on the specified column, with the given pattern head.

compile :: (MonadMatch m, Cored c) => Matching c a -> PatternMatrix t c a -> m (c a)

Compiles a pattern matrix together with a corresponding set of core branches to a final Core value, which will be the decision tree version of the pattern matrix.

compileBinding :: (MonadMatch m, Cored c) => [[Pattern t]] -> [Guarded (Scope BodyBound c a)] -> [[Scope (Var WhereBound Word32) c a]] -> m (Scope Word32 c a)

compileLambda :: (MonadMatch m, Cored c) => [Pattern t] -> Scope PatternPath c a -> m (Scope Word8 c a)

compileCase :: (MonadMatch m, Cored c) => [Pattern t] -> c a -> [Guarded (Scope PatternPath c a)] -> m (c a)