Ermine.Meta

Description

Meta variables for unification

Documentation

data Meta s f

A meta variable for skolemization and unification

Constructors

Meta !Int (STRef s (Maybe (f (Meta s f))))

Our unification monad

data U s a

U provides a fresh variable supply for meta variables.

evalU :: (forall s. U s a) -> a

Evaluate an expression in the U monad with a fresh variable supply.

newMeta :: U s (Meta s f)

Construct a new meta variable

readMeta :: Meta s f -> U s (Maybe (f (Meta s f)))

Read a meta variable

writeMeta :: Meta s f -> f (Meta s f) -> U s ()

Write to a meta variable

zonkMeta :: (Monad f, Traversable f) => f (Meta s f) -> U s (f (Meta s f))

Expand meta variables recursively