Ermine.Parser.Global

Description

This module provides a parser for global names.

Note: most parsing goes through strings, which are later fixed up into global names, but in some cases globals should be parsed directly.

Documentation

globalIdent :: (Monad m, TokenParsing m) => IdentifierStyle m -> m Global

Parse a global identifier with the given style. TODO: package/module information