Ermine.Parser.Type

Description

This module provides the parser for types

Documentation

type Ann = Annot (Maybe Text) Text

type Typ = Type (Maybe Text) Text

anyType :: Ann

typ :: (Monad m, TokenParsing m) => m Typ

Parse a Type.

typ0 :: (Monad m, TokenParsing m) => m Typ

Parse an atomic type

annotation :: (Monad m, TokenParsing m) => m Ann

typVarBinding :: (Monad m, TokenParsing m) => m ([Text], Kind (Maybe Text))

Parses an optionally annotated type variable.

typVarBinding ::= ( ident : kind ) | ident

typVarBindings :: (Monad m, TokenParsing m) => m [(Text, Kind (Maybe Text))]

Parses a series of type var bindings, processing the result to a more usable format.

typVarBindings ::= emtpy | someTypVarBindings

someTypVarBindings :: (Monad m, TokenParsing m) => m [(Text, Kind (Maybe Text))]

Parses a series of type var bindings, processing the result to a more usable format.

someTypVarBindings ::= typVarBinding typVarBindings

quantifier :: (Monad m, TokenParsing m) => String -> m ([Text], [(Text, Kind (Maybe Text))])

Parses a quantifier.

quantifier ::= Q quantBindings .

quantBindings :: (Monad m, TokenParsing m) => m ([Text], [(Text, Kind (Maybe Text))])

Parses the bound variables for a quantifier.

quantBindings ::= {kindVars} | someTypVarBindings | {kindVars} typVarBindings