Ermine.Pattern.Env

Description

Documentation

newtype PatternEnv

The environment necessary to perform pattern compilation. We need two pieces of information:

1) The signatures (full list of constructors) associated with each particular constructor

2) A mapping from constructors to their Core integer tag

This is accomplished via a map of maps. The outer map should take each global to its associated signature, and signatures are represented as maps from globals to integer tags.

Constructors

PatternEnv

Fields

signatures :: HashMap Global (HashMap Global Word8)

dummyPatternEnv :: PatternEnv

class (Applicative m, Monad m) => MonadPattern m where

Monads that allow us to perform pattern compilation, by providing a PatternEnv.

Methods

askPattern :: m PatternEnv

isSignature :: MonadPattern m => Set PatternHead -> m Bool

Determines whether a set of pattern heads constitutes a signature. This is handled specially for tuples and literals, and relies on the monad for data type constructors.

constructorTag :: MonadPattern m => PatternHead -> m Word8

Looks up the constructor tag for a pattern head. For tuples this is always 0, but constructors must consult the compilation environment.

constructorGlobal :: PatternHead -> Global

Looks up the constructor tag for a pattern head. For tuples this is always 0, but constructors must consult the compilation environment.