Ermine.Pattern.Matching

Description

Matching

data Matching c a

Additional information needed for pattern compilation that does not really belong in the pattern matrix.

The pathMap stores information for resolving previous levels of the pattern to appropriate cores for the current context.

colCores contains core terms that refer to each of the columns of the matrix in the current context.

colPaths contains the paths into the original pattern that correspond to each of the current columns.

Constructors

Matching

Fields

_pathMap :: HashMap PatternPath (c a)

_colCores :: [c a]

_colPaths :: [PatternPaths]

class HasMatching t c a | t -> c a where

Minimal complete definition

matching

Methods

matching :: Lens' t (Matching c a)

colCores :: Lens' t [c a]

colPaths :: Lens' t [PatternPaths]

pathMap :: Lens' t (HashMap PatternPath (c a))

Compilation

compile :: (MonadPattern m, Cored c) => Matching c a -> PatternMatrix t c a -> m (c a)

Compiles a pattern matrix together with a corresponding set of core branches to a final Core value, which will be the decision tree version of the pattern matrix.

compileBinding :: (MonadPattern m, Cored c) => [[Pattern t]] -> [Guarded (Scope BodyBound c a)] -> [[Scope (Var WhereBound Word32) c a]] -> m (Scope Word32 c a)

compileLambda :: (MonadPattern m, Cored c) => [Pattern t] -> Scope PatternPath c a -> m (Scope Word8 c a)

compileCase :: (MonadPattern m, Cored c) => [Pattern t] -> c a -> [Guarded (Scope PatternPath c a)] -> m (c a)