Ermine.Pattern.Matrix

Description

Pattern Matrix

data PatternMatrix t c a

Pattern matrices for compilation. The matrix is represented as a list of columns. There is also an extra column representing the guards.

Constructors

PatternMatrix

Fields

_cols :: [[Pattern t]]

_bodies :: [Claused c a]

class HasPatternMatrix t t c a | t -> t c a where

Minimal complete definition

patternMatrix

Methods

patternMatrix :: Lens' t (PatternMatrix t c a)

bodies :: Lens' t [Claused c a]

cols :: Lens' t [[Pattern t]]

defaultOn :: Applicative c => Int -> PatternMatrix t c a -> PatternMatrix t c (Var () (c a))

Computes the matrix that should be used recursively when defaulting on the specified column.

splitOn :: Applicative c => Int -> PatternHead -> PatternMatrix t c a -> PatternMatrix t c (Var Word64 (c a))

Computes the matrix that should be used recursively when defaulting on the specified column, with the given pattern head.

Claused

data Claused c a

Constructors

Localized [Scope PatternPath (Scope Word64 c) a] (Guarded (Scope PatternPath (Scope Word64 c) a))

Raw (Guarded (Scope PatternPath c a))

hoistClaused :: Functor c => (forall x. c x -> d x) -> Claused c a -> Claused d a