Ermine.Pretty.Kind

Description

Pretty Printing Kinds

prettyHardKind :: HardKind -> Doc

prettyKind :: Kind String -> Bool -> Doc

Pretty print a Kind

prettySchema :: Schema String -> [String] -> Doc

Pretty print a Kind, using a fresh kind variable supply

You should have already removed any free variables from the variable set.