Ermine.Pretty.Type

Description

Pretty Printing Kinds

prettyHardType :: HardType -> Doc

prettyType :: Type String String -> [String] -> Int -> Doc

Pretty print a Type using a fresh variable supply and an ambient precedence.

You should have already removed any free variables from the variable set.

prettyTypeSchema :: Scope Int (TK String) String -> ([Hint], [Hint]) -> [String] -> Doc