Ermine.Scope

Description

Documentation

hoistScope :: Functor f => (forall x. f x -> g x) -> Scope b f a -> Scope b g a

Lift a natural transformation from f to g into one between scopes.

bitraverseScope :: (Bitraversable t, Applicative f) => (k -> f k') -> (a -> f a') -> Scope b (t k) a -> f (Scope b (t k') a')

This allows you to bitraverse a Scope.

bound :: Prism (Var a c) (Var b c) a b

This prism acts like a reversible smart constructor for B.

free :: Prism (Var c a) (Var c b) a b

class Monad m => BoundBy tm m | tm -> m where

Generalizes Bound to permit binding by another type without taking it as a parameter.

Methods

boundBy :: (a -> m b) -> tm a -> tm b