Ermine.Sharing

Description

These combinators can be used to retain sharing information.

Documentation

runSharing :: Monad m => a -> WriterT Any m a -> m a

Run an action, if it returns Any True then use its new value, otherwise use the passed in value.

This can be used to recover sharing during unification when no interesting unification takes place.

This version discards the Writer wrapper.

sharing :: MonadWriter Any m => a -> m a -> m a

Run an action, if it returns Any True then use its new value, otherwise use the passed in value.

This can be used to recover sharing during unification when no interesting unification takes place.

This version retains the Writer monad wrapper.