Ermine.Syntax

Description

Variables

class Variable t where

Discriminable pure.

Methods

var :: Prism' (t a) a

If t is also Applicative, then it should satisfy these laws:

 pure a ^? var ≡ Just a
 pure ≡ review var

Application

class App t where

Discriminable App

This prism provides ad hoc overloading of construction and pattern matching on App.

Methods

app :: Prism' t (t, t)

apps :: App t => t -> [t] -> t

Fold a series of applications.

(#) :: AReview s t a b -> b -> t

An infix alias for review.

 unto f # x ≡ f x
 l # x ≡ x ^. re l

This is commonly used when using a Prism as a smart constructor.

>>> _Left # 4
Left 4

But it can be used for any Prism

>>> base 16 # 123
"7b"

 (#) :: Iso'      s a -> a -> s
 (#) :: Prism'    s a -> a -> s
 (#) :: Review'   s a -> a -> s
 (#) :: Equality' s a -> a -> s

Fun

class Fun t where

Discriminable syntactic arrows.

Methods

fun :: Prism' t (t, t)

(~>) :: Fun t => t -> t -> t

Provide ad hoc overloading of function arrows.

Tup

class (Reviewable p, Functor f) => Tup p f t where

Irreversible tupling.

(Possibly) Discriminable tupling.

Methods

tupled :: Overloaded' p f t [t]

tup :: Tup Reviewed Identity t => [t] -> t