Ermine.Syntax.Core

Description

Core Terms

data Core a

Core values are the output of the compilation process.

They are terms where the dictionary passing has been made explicit and all of the types have been checked and removed.

Constructors

Var a

HardCore !HardCore

Data !Int [Core a]

App !(Core a) !(Core a)

Lam !Int !(Scope Int Core a)

Let [Scope Int Core a] !(Scope Int Core a)

Case !(Core a) (Map Int (Int, Scope Int Core a)) (Maybe (Scope () Core a))

Dict

Fields

supers :: [Core a]

slots :: [Scope Int Core a]

LamDict !(Scope () Core a)

AppDict !(Core a) !(Core a)

class AsAppDict c where

Methods

_AppDict :: Prism' c (c, c)

appDicts :: AsAppDict c => c -> [c] -> c

data HardCore

Constructors

Super !Int

Slot !Int

Lit !Literal

Error !String

class AsHardCore c where

Methods

_HardCore :: Prism' c HardCore

_Lit :: Prism' c Literal

_Error :: Prism' c String

_Super :: Prism' c Int

_Slot :: Prism' c Int

class Lit a where

Lifting of literal values to core.

Methods

lit :: a -> Core b

lits :: [a] -> Core b

super :: Int -> Core a -> Core a

slot :: Int -> Core a -> Core a

Smart constructors

lam :: Eq a => [a] -> Core a -> Core a

smart lam constructor

let_ :: Eq a => [(a, Core a)] -> Core a -> Core a

smart let constructor

dataCon :: Int -> Int -> Core a

Builds an n-ary data constructor

Common built-in terms

cons :: Cons Reviewed Identity s s a a => a -> s -> s

cons an element onto a container.

nil :: Core a

The built-in '[]' constructor for a list.

just :: Core a -> Core a

The built-in Just constructor for Maybe.

nothing :: Core a

The built-in Nothing constructor for Maybe.