Ermine.Syntax.Digest

Description

Constructing Digests

class Digestable t where

Methods

digest :: Hash ctx d => ctx -> t -> ctx

update a hash. This will default to updating using show if not specified

class DigestableList t where

A helper class provided to aid in hash calculation

Methods

digestList :: Hash ctx d => ctx -> [t] -> ctx