Ermine.Syntax.Global

Description

Documentation

data Global

A Global is a full qualified top level name.

NB: You should construct these with global and only use the constructor for pattern matching.

Constructors

Global !ByteString !Fixity !ModuleName !Text

class HasGlobal t where

Methods

global :: Lens' t Global

fixity :: Lens' t Fixity

A lens that will read or update the fixity (and compute a new digest)

class AsGlobal t where

Methods

_Global :: Prism' t Global

glob :: AsGlobal t => Fixity -> ModuleName -> Text -> t

Construct a Global with a correct digest.

data Assoc

The associativity of an infix identifier

Constructors

L

R

N

data Fixity

The fixity of an identifier

Constructors

Infix !Assoc !Int

Prefix !Int

Postfix !Int

Idfix

_Fixity :: Prism' Word8 Fixity

Lenses

unpackFixity :: Monad m => Word8 -> m Fixity

packFixity :: Fixity -> Word8

Pack Fixity into a Word8.

Format:

 01234567
 ccaapppp

cc is constructor tag, 0-3 pppp is precedence level, 0-9 aa is associativity tag, 0-2