Ermine.Syntax.Kind

Description

Kinds

data Kind a

Kinds with kind variables of type a

Constructors

Var a

(Kind a) :-> (Kind a)

HardKind HardKind

Hard Kinds

data HardKind

Kinds that can be compared by simple structural equality

Constructors

Star

Constraint

Rho

Phi

class Kindly k where

Smart constructor for hard kinds

Methods

hardKind :: Prism' k HardKind

star, phi, rho, constraint :: k

Kind Schemas

data Schema a

Kind schemas

Constructors

Schema !Int !(Scope Int Kind a)

schema :: Kind a -> Schema a

Lift a kind into a kind schema

>>> schema (star ~> star)
Schema 0 (Scope (Var (F (HardKind Star :-> HardKind Star))))

general :: Ord k => Kind k -> Schema a

Construct a schema from a kind, generalizing all free variables.

When working with Meta variables, you want to use generalize to zonk the kind and check skolems.

>>> general "a"
Schema 1 (Scope (Var (B 0)))

>>> general ("a" ~> "a")
Schema 1 (Scope (Var (B 0) :-> Var (B 0)))

>>> general ("a" ~> "b")
Schema 2 (Scope (Var (B 0) :-> Var (B 1)))

>>> general ("b" ~> "a")
Schema 2 (Scope (Var (B 0) :-> Var (B 1)))

>>> general (star ~> star)
Schema 0 (Scope (HardKind Star :-> HardKind Star))

Kind Variables

class HasKindVars s t a b | s -> a, s b -> t, t -> b, t a -> s where

Provides a traversal of free kind variables that can be used to perform substitution or extract a free variable set.

Methods

kindVars :: Traversal s t a b

Pretty Printing

prettyKind :: Applicative f => Kind a -> Bool -> (a -> Bool -> f (Doc b)) -> f (Doc b)

Pretty print a Kind, using a helper to print free variables

prettySchema :: Applicative f => Schema a -> [String] -> (a -> Bool -> f (Doc b)) -> f (Doc b)

Pretty print a Kind, using a fresh kind variable supply and a helper to print free variables

You should have already removed any free variables from the variable set.