Ermine.Syntax.Scope

Description

This module provides extensions to the bound package.

Documentation

hoistScope :: Functor f => (forall x. f x -> g x) -> Scope b f a -> Scope b g a

Lift a natural transformation from f to g into one between scopes.

bitraverseScope :: (Bitraversable t, Applicative f) => (k -> f k') -> (a -> f a') -> Scope b (t k) a -> f (Scope b (t k') a')

This allows you to bitraverse a Scope.

transverseScope :: (Applicative f, Monad f, Traversable g) => (forall r. g r -> f (h r)) -> Scope b g a -> f (Scope b h a)

This is a higher-order analogue of traverse.

class Monad m => BoundBy tm m | tm -> m where

Generalizes Bound to permit binding by another type without taking it as a parameter.

Methods

boundBy :: (a -> m b) -> tm a -> tm b

instantiateVars :: Monad t => [a] -> Scope Int t a -> t a

instantiate bound variables using a list of new variables

serializeScope3 :: MonadPut m => (b -> m ()) -> (forall a. (a -> m ()) -> f a -> m ()) -> (v -> m ()) -> Scope b f v -> m ()

deserializeScope3 :: MonadGet m => m b -> (forall a. m a -> m (f a)) -> m v -> m (Scope b f v)