Ermine.Syntax.Term

Description

This module provides the AST for Terms

Terms

data Term t a

Terms in the Ermine language.

Constructors

Var a

App !(Term t a) !(Term t a)

HardTerm !HardTerm

Sig !(Term t a) t

Lam [Pattern t] !(Scope PatPath (Term t) a)

Case !(Term t a) [Alt t (Term t) a]

Let [Binding t a] !(Scope Int (Term t) a)

Loc !Rendering !(Term t a)
informational link to the location the term came from

Remember !Int !(Term t a)
Used to provide hole support.

bindTerm :: (t -> t') -> (a -> Term t' b) -> Term t a -> Term t' b

Perform simultaneous substitution on terms and type annotations.

Hard Terms

data HardTerm

Simple terms that can be compared with structural equality.

Constructors

Lit Literal

DataCon !Global (Type Void Void)

Tuple !Word8

Hole
A placeholder that can take any type. Easy to Remember.

class Terminal t where

This class provides a prism to match against or inject a HardTerm.

Methods

hardTerm :: Prism' t HardTerm

litTerm :: Literal -> t

hole :: t

Bindings

data DeclBound

Bound variables in a declaration are rather complicated. One can refer to any of the following: 1. Definitions in the same declaration sequence 2. Variables bound in a pattern 3. Definitions in a where clause the DeclBound type captures these three cases in the respective constructors.

Constructors

D Int

P PatPath

W Int

data Binding t a

A Binding provides its source location as a rendering, knowledge of if it is explicit or implicitly bound and a list of right hand side bindings.

Constructors

Binding !Rendering !(BindingType t) [Body t a]

data BindingType t

Indicate if a definition is explicitly bound with a type annotation or implicitly bound without.

Constructors

Explicit t

Implicit

data Body t a

A body is the right hand side of a definition. This isn't a term because it has to perform simultaneous matches on multiple patterns with backtracking. Each Body contains a list of where clause bindings to which the body and guards can refer.

Constructors

Body [Pattern t] (Guarded (Scope DeclBound (Term t) a)) [Binding t (Var PatPath a)]