Ermine.Syntax.Type

Description

Types

data Type k a

Ermine types, parameterized by their free kind variables and free type variables

Constructors

Var a

App !(Type k a) !(Type k a)

HardType !HardType

Forall !Int [Scope Int Kind k] [Scope Int (TK k) a] !(Scope Int (TK k) a)

Loc !Rendering !(Type k a)

Exists [Kind k] [Scope Int (Type k) a]

type FieldName = String

A placeholder for a more complicated database fieldname.

Hard Types

data HardType

A Type that can be compared with mere structural equality.

Constructors

Tuple !Int

Arrow

Con !Global !(Schema Void)

ConcreteRho !(Set FieldName)

class Typical t where

Smart constructors that allows us to pun various HardType constructor names for Type.

Methods

hardType :: Prism' t HardType

tuple :: Int -> t

arrow :: t

con :: Global -> Schema Void -> t

concreteRho :: Set FieldName -> t

Binding

newtype TK k a

TK is a special version of Scope for types that binds kinds. It is used by Forall.

Constructors

TK

Fields

runTK :: Type (Var Int (Kind k)) a

abstractKinds :: (k -> Maybe Int) -> Type k a -> TK k a

Bind some of the kinds referenced by a Type.

instantiateKinds :: (Int -> Kind k) -> TK k a -> Type k a

Instantiate the kinds bound by a TK obtaining a Type.

instantiateKindVars :: [k] -> TK k a -> Type k a

Instantiate kinds using a list of variables.

bindType :: (k -> Kind k') -> (a -> Type k' b) -> Type k a -> Type k' b

Perform simultaneous substitution on kinds and types in a Type.

Type Variables

class HasTypeVars s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide a Traversal of type variables that can be used to extract them or substitute them for other type variables.

Methods

typeVars :: Traversal s t a b