Ermine.Term

Description

Terms

data Term t a

Terms in the Ermine language.

Constructors

Var a

App !(Term t a) !(Term t a)

HardTerm !HardTerm

Sig !(Term t a) t

Lam !(Pat t) !(Scope Int (Term t) a)

Case !(Term t a) [Alt t (Term t) a]

Let [Binding t a] !(Scope Int (Term t) a)

Loc !Rendering !(Term t a)
informational link to where the term came from

Remember !Int !(Term t a)
Used to provide hole support.

bindTerm :: (t -> t') -> (a -> Term t' b) -> Term t a -> Term t' b

Perform simultaneous substitution on terms and type annotations.

Hard Terms

data HardTerm

Simple terms that can be compared with structural equality.

Constructors

Prim Prim

Hole
A placeholder that can take any type. Easy to Remember.

class Terminal t where

Methods

hardTerm :: Prism' t HardTerm

primTerm :: Prim -> t

hole :: t

Bindings

data Binding t a

A Binding provides its source location as a rendering, knowledge of if it is explicit or implicitly bound and a list of right hand side bindings.

Constructors

Binding !Rendering !(BindingType t) [Body t a]

data BindingType t

Indicate if a definition is explicitly bound with a type annotation or implicitly bound without.

Constructors

Explicit t

Implicit

data Body t a

A body is the right hand side of a definition. This isn't a term because it has to perform simultaneous matches on multiple patterns with backtracking.

Constructors

Body [Pat t] !(Scope (Either Int Int) (Term t) a)