Ermine.Unification

Description

Unification support.

Our unification monad

data U s a

The unification monad provides a fresh variable supply and tracks a current Rendering to blame for any unification errors.

evalU :: Rendering -> (forall s. U s a) -> Either (Rendering, String) a

Evaluate an expression in the U Monad with a fresh variable supply.

fresh :: U s Int

Generate a fresh variable

st :: ST s a -> U s a

Lift an ST action into the U monad.