Ermine.Unification.Kind

Description

Documentation

type MetaK s = Meta s Kind ()

A kind meta-variable

type KindM s = Kind (MetaK s)

A kind filled with meta-variables

unifyKind :: (MonadMeta s m, MonadWriter Any m) => IntSet -> KindM s -> KindM s -> m (KindM s)

Returns the a unified form if different from the left argument.

The writer is used to track if any interesting edits have been made

unifyKindVar :: (MonadMeta s m, MonadWriter Any m) => IntSet -> MetaK s -> KindM s -> m (KindM s)

Unify a known Meta variable with a kind that isn't a Var.

kindOccurs :: MonadMeta s m => Set (MetaK s) -> m a

Die with an error message due to a cycle between the specified kinds.

>>> test $ do k <- Var <$> newMeta (); unifyKind mempty k (star ~> k)
*** Exception: (interactive):1:1: error: infinite kind detected
cyclic kind: a = * -> a

>>> test $ do k1 <- Var <$> newMeta (); k2 <- Var <$> newMeta (); unifyKind mempty k1 (k1 ~> k2); unifyKind mempty k2 (k1 ~> k2); return (k1 ~> k2)
*** Exception: (interactive):1:1: error: infinite kinds detected in (a -> b)
cyclic kind: a = a -> b
cyclic kind: b = a -> b

generalize :: MonadMeta s m => KindM s -> m (Schema a)

Generalize a Kind, checking for escaped Skolems.