Ermine.Unification.Meta

Description

Meta variables

Meta variables

data Meta s f a

A meta variable for skolemization and unification

Constructors

Meta a Hint !Int !(STRef s (Maybe (f (Meta s f a)))) !(STRef s Depth) !(STRef s Rank)

metaId :: forall s f a. Lens' (Meta s f a) Int

metaValue :: forall s f a. Lens' (Meta s f a) a

metaRef :: forall s f a f. Lens (Meta s f a) (Meta s f a) (STRef s (Maybe (f (Meta s f a)))) (STRef s (Maybe (f (Meta s f a))))

metaHint :: forall s f a. Lens' (Meta s f a) Hint

Meta Types

type MetaT s = Meta s (Type (MetaK s)) (KindM s)

A type meta-variable

type TypeM s = Type (MetaK s) (MetaT s)

A type filled with meta-variables

Meta Kinds

type MetaK s = Meta s Kind Bool

A kind meta-variable

type KindM s = Kind (MetaK s)

A kind filled with meta-variables

λ-Depth

type Depth = Int

Kuan and MacQueen's notion of λ-ranking.

metaDepth :: forall s f a. Lens' (Meta s f a) (STRef s Depth)

depthInf :: Depth

This plays the role of infinity.

Union-By-Rank

type Rank = Word

Rank for union-by-rank viewing meta-variables as disjoint set forests

metaRank :: forall s f a. Lens' (Meta s f a) (STRef s Rank)

bumpRank :: STRef s Rank -> ST s ()

Working with Meta

newMeta :: MonadMeta s m => a -> Hint -> m (Meta s f a)

Construct a new meta variable

newShallowMeta :: MonadMeta s m => Depth -> a -> Hint -> m (Meta s f a)

Construct a new meta variable at a given depth

readMeta :: MonadMeta s m => Meta s f a -> m (Maybe (f (Meta s f a)))

Read a meta variable

writeMeta :: MonadMeta s m => Meta s f a -> f (Meta s f a) -> m ()

Write to a meta variable

Pruning

semiprune :: (Variable f, Monad f, MonadMeta s m) => f (Meta s f a) -> m (f (Meta s f a))

Path-compression

zonk :: (MonadMeta s m, MonadWriter Any m, Traversable f, Monad f) => f (Meta s f a) -> m (f (Meta s f a))

Expand meta variables recursively

zonk_ :: (MonadMeta s m, Traversable f, Monad f) => f (Meta s f a) -> m (f (Meta s f a))

zonkWith :: (MonadMeta s m, MonadWriter Any m, Traversable f, Monad f) => f (Meta s f a) -> (Meta s f a -> m ()) -> m (f (Meta s f a))

zonkScope :: (MonadMeta s m, MonadWriter Any m, Traversable f, Monad f) => Scope b f (Meta s f a) -> m (Scope b f (Meta s f a))

zonkScope_ :: (MonadMeta s m, Traversable f, Monad f) => Scope b f (Meta s f a) -> m (Scope b f (Meta s f a))

hasSkolems :: (Traversable f, Monad f, MonadMeta s m) => LensLike' m ts (f (Meta s f a)) -> [Meta s f a] -> ts -> m ts

Checks if any of the listed skolems exists in the types given in the traversal.

checkSkolems :: MonadMeta s m => Maybe Depth -> LensLike' m ts (TypeM s) -> [MetaT s] -> ts -> m ts

Check for escaped Skolem variables in any container full of types.

Dies due to an escaped Skolem or returns the container with all of its types fully zonked.

checkEscapes :: MonadMeta s m => Depth -> [Meta s f a] -> m ()

Checks if any of the skolems in the list escapes into the context defined by the given depth.

checkDistinct :: (Traversable f, Monad f, Variable f, MonadMeta s m) => [Meta s f a] -> m [Meta s f a]

MetaEnv

data MetaEnv s

class HasMetaEnv t s | t -> s where

Minimal complete definition

metaEnv

Methods

metaEnv :: Lens' t (MetaEnv s)

metaFresh :: Lens' t (STRef s Int)

metaMemory :: Lens' t (STRef s (IntMap (TypeM s)))

metaRendering :: Lens' t Rendering

class (Applicative m, MonadST m, s ~ World m) => MonadMeta s m | m -> s where

Minimal complete definition

localMeta

Methods

askMeta :: m (MetaEnv s)

localMeta :: (MetaEnv s -> MetaEnv s) -> m a -> m a

viewMeta :: MonadMeta s m => Getting a (MetaEnv s) a -> m a

The unification monad

data M s a

The unification monad provides a fresh variable supply and tracks a current Rendering to blame for any unification errors.

runM :: Rendering -> (forall s. M s a) -> Either Diagnostic a

Evaluate an expression in the M Monad with a fresh variable supply.

runM_ :: Rendering -> (forall s. M s a) -> a

Evaluate an expression in the M Monad with a fresh variable supply, throwing any errors returned.

ioM :: MonadIO m => Rendering -> (forall s. M s a) -> m a

Evaluate an expression in the M Monad with a fresh variable supply, throwing any errors returned.

throwM :: MonadST m => Diagnostic -> m a

Throw a Diagnostic error.

fresh :: MonadMeta s m => m Int

Generate a fresh variable

remember :: MonadMeta s m => Int -> TypeM s -> m ()