Ermine.Unification.Sharing

Description

These combinators can be used to retain sharing information.

Documentation

runSharing :: Monad m => a -> SharingT m a -> m a

Run an action, if it returns Any True then use its new value, otherwise use the passed in value.

This can be used to recover sharing during unification when no interesting unification takes place.

This version discards the SharingT wrapper.

sharing :: MonadWriter Any m => a -> m a -> m a

Run an action, if it returns Any True then use its new value, otherwise use the passed in value.

This can be used to recover sharing during unification when no interesting unification takes place.

This version retains the current monad wrapper.

newtype SharingT m a

Constructors

SharingT

Fields

unsharingT :: m (Shared a)

data Shared a

Constructors

Shared !Bool a