Ermine.Unification.Type

Description

This module tells us how to unify types. We use a simplified form of HMF-style unification to deal with unification of foralls.

Documentation

type MetaT s = Meta s (Type (MetaK s)) (KindM s)

A type meta-variable

type TypeM s = Type (MetaK s) (MetaT s)

A type filled with meta-variables

unifyType :: (MonadWriter Any m, MonadMeta s m) => IntSet -> TypeM s -> TypeM s -> m (TypeM s)

Unify two types, with access to a visited set, logging to a Writer whether or not the answer differs from the first type argument.