Ermine.Diagnostic

Description

This module re-exports Rendering from trifecta along with our in-house diagnostic type.

This allows the non-parsing modules to not have to concern themselves with how we choose to represent source locations.

Rendering

data Rendering

class HasRendering t where

Methods

rendering :: Lens t t Rendering Rendering

renderingDelta :: Lens t t Delta Delta

renderingLine :: Lens t t (Lines -> Lines) (Lines -> Lines)

renderingLineBytes :: Lens t t Int64 Int64

renderingLineLen :: Lens t t Int64 Int64

renderingOverlays :: Lens t t (Delta -> Lines -> Lines) (Delta -> Lines -> Lines)

Diagnostics

data Diagnostic

Ermine diagnostic type

Constructors

Diagnostic !Rendering (Maybe TermDoc) [TermDoc] (Set String)

class AsDiagnostic t where

This provides the _Diagnostic prism that can be used when matching against SomeException or in a custom Error type.

Methods

_Diagnostic :: Prism' t Diagnostic

data Err

Constructors

Err (Maybe TermDoc) [TermDoc] (Set String)

class HasErr t where

Methods

err :: Lens t t Err Err

expected :: Lens t t (Set String) (Set String)

footnotes :: Lens t t [TermDoc] [TermDoc]

reason :: Lens t t (Maybe TermDoc) (Maybe TermDoc)

die :: HasRendering r => r -> String -> Diagnostic

Construct a diagnostic