Ermine.Parser.Keywords

Description

Documentation

startingKeywords :: HashSet String

This is the set of keywords that can only occur at the beginning of the line for auto-completion purposes.

otherKeywords :: HashSet String

This is the set of keywords that can occur anywhere on the line for auto-completion purposes.

keywords :: HashSet String

The set of all keywords.

keywords = startingKeywords <> otherKeywords