Ermine.Pretty.Kind

Description

Pretty Printing Kinds

prettyHardKind :: HardKind -> Doc a

prettyKind :: Applicative f => Kind a -> Bool -> (a -> Bool -> f (Doc b)) -> f (Doc b)

Pretty print a Kind, using a helper to print free variables

prettyKind_ :: Kind String -> Doc a

prettySchema :: Applicative f => Schema a -> [String] -> (a -> Bool -> f (Doc b)) -> f (Doc b)

Pretty print a Kind, using a fresh kind variable supply and a helper to print free variables

You should have already removed any free variables from the variable set.