Ermine.Syntax

Description

Variables

class Variable t where

Discriminable pure.

Minimal complete definition

Nothing

Methods

_Var :: Prism' (t a) a

If t is also Applicative, then it should satisfy these laws:

pure a '^?' _Var ≡ Just a
pure ≡ review _Var

Application

class App t where

Discriminable App

This prism provides ad hoc overloading of construction and pattern matching on App.

Minimal complete definition

Nothing

Methods

_App :: Prism' (t a) (t a, t a)

apps :: App t => t a -> [t a] -> t a

Fold a series of applications.

(##) :: App t => t a -> t a -> t a infixl 9

Convenient infix application operator.

Strict Unboxed Application

class AppHash t where

Minimal complete definition

Nothing

Methods

_AppHash :: Prism' (t a) (t a, t a)

appHashes :: AppHash t => t a -> [t a] -> t a

Fold a series of unboxed applications.

Dictionary Application

class AppDict c where

AppDict provides strict application for dictionary manipulation

Minimal complete definition

Nothing

Methods

_AppDict :: Prism' (c a) (c a, c a)

appDicts :: AppDict t => t a -> [t a] -> t a

Apply several dictionaries.

Fun

class Fun t where

Discriminable syntactic arrows.

Minimal complete definition

Nothing

Methods

_Fun :: Prism' (t a) (t a, t a)

(~>) :: Fun t => t a -> t a -> t a infixr 0

Provide ad hoc overloading of function arrows.

Tup

class (Reviewable p, Functor f) => Tup p f t where

(Possibly) Discriminable tupling.

Minimal complete definition

Nothing

Methods

_Tup :: Optic' p f t [t]

tup :: Tup Tagged Identity t => [t] -> t

tup' :: Tup Tagged Identity t => [t] -> t