Ermine.Syntax

Description

Variables

class Variable t where

Discriminable pure.

Methods

var :: Prism' (t a) a

If t is also Applicative, then it should satisfy these laws:

 pure a ^? var ≡ Just a
 pure ≡ review var

Application

class App t where

Discriminable App

This prism provides ad hoc overloading of construction and pattern matching on App.

Methods

app :: Prism' (t a) (t a, t a)

apps :: App t => t a -> [t a] -> t a

Fold a series of applications.

(#) :: App t => t a -> t a -> t a

Convenient infix application operator.

Fun

class Fun t where

Discriminable syntactic arrows.

Methods

fun :: Prism' (t a) (t a, t a)

(~>) :: Fun t => t a -> t a -> t a

Provide ad hoc overloading of function arrows.