Ermine.Syntax.Core

Description

Core Terms

data Core a

Core values are the output of the compilation process.

They are terms where the dictionary passing has been made explicit and all of the types have been checked and removed.

Constructors

Var a

HardCore !HardCore

Data !Word8 [Core a]

App !(Core a) !(Core a)

Lam !Word8 !(Scope Word8 Core a)

Let [Scope Word32 Core a] !(Scope Word32 Core a)

Case !(Core a) (Map Word8 (Word8, Scope Word8 Core a)) (Maybe (Scope () Core a))

Dict

Fields

supers :: [Core a]

slots :: [Scope Word8 Core a]

LamDict !(Scope () Core a)

AppDict !(Core a) !(Core a)

class (Applicative c, Monad c) => Cored c where

Instances of this are things we can both construct from a Core expression, and perform case analysis upon.

e.g. Core and Scope b Core

Minimal complete definition

core, caze, lambda, lambdaDict, letrec

Methods

core :: Core a -> c a

caze :: c a -> Map Word8 (Word8, Scope Word8 c a) -> Maybe (Scope () c a) -> c a

lambda :: Word8 -> Scope Word8 c a -> c a

lambdaDict :: Scope () c a -> c a

letrec :: [Scope Word32 c a] -> Scope Word32 c a -> c a

hardCore :: HardCore -> c a

class AsAppDict c where

AppDict provides strict application for dictionary manipulation

Minimal complete definition

Nothing

Methods

_AppDict :: Prism' c (c, c)

appDicts :: AsAppDict c => c -> [c] -> c

Apply several dictionaries.

data JavaLike

Constructors

Method !Bool !Text !Text [Text]
Java methods: static, class name, method name, arg class names

Constructor !Text [Text]
Java constructors: class name, arg class names

Value !Bool !Text !Text
Java values: static, class name, field name

data Foreign

Constructors

JavaLike !JavaLike

Unknown !Text

data HardCore

HardCore is the subset of Core terms that can be unified with value equality.

Constructors

Super !Word8

Slot !Word8

Lit !Literal

PrimOp !Text

Foreign !Foreign

Error !Text

GlobalId !Global

InstanceId !Head

class AsHardCore c where

This class describes things that could be HardCore.

Minimal complete definition

Nothing

Methods

_HardCore :: Prism' c HardCore

_Lit :: Prism' c Literal

_PrimOp :: Prism' c Text

_Error :: Prism' c Text

_Super :: Prism' c Word8

_Slot :: Prism' c Word8

_Foreign :: Prism' c Foreign

_GlobalId :: Prism' c Global

_InstanceId :: Prism' c Head

class Lit a where

Lifting of literal values to core.

Minimal complete definition

lit

Methods

lit :: a -> Core b

lits :: [a] -> Core b

super :: (AsHardCore c, AsAppDict c) => Word8 -> c -> c

ask for the nth the superclass of a given dictionary as a core expression

slot :: (AsHardCore c, AsAppDict c) => Word8 -> c -> c

ask for the nth slot of a given dictionary as a core expression

Smart constructors

lam :: (Cored m, Eq a) => [a] -> Core a -> m a

Smart Lam constructor

let_ :: (Cored m, Eq a) => [(a, Core a)] -> Core a -> m a

Smart Let constructor

dataCon :: Cored m => Word8 -> Word8 -> m a

Builds an n-ary data constructor

Common built-in terms

cons :: Cons s s a a => a -> s -> s

nil :: Core a

The built-in '[]' constructor for a list.

just :: Core a -> Core a

The built-in Just constructor for Maybe.

nothing :: Core a

The built-in Nothing constructor for Maybe.