Ermine.Syntax.Core

Description

Core Terms

data Core a

Core values are the output of the compilation process.

They are terms where the dictionary passing has been made explicit and all of the types have been checked and removed.

Constructors

Var a

Prim !Prim [Core a]

App !(Core a) !(Core a)

Lam !(Pat ()) !(Scope Int Core a)

Let [Scope Int Core a] !(Scope Int Core a)

Case !(Core a) [Alt () Core a]

class Lit a where

Lifting of literal values to core.

Methods

lit :: a -> Core b

lits :: [a] -> Core b

Smart patterns

data P a

Smart Pattern

varp :: a -> P a

A pattern that binds a variable.

_p :: P a

A wildcard pattern that ignores its argument

strictp :: P a -> P a

A strict (bang) pattern

lazyp :: P a -> P a

A lazy (irrefutable) pattern

asp :: a -> P a -> P a

An as (@) pattern.

primp :: Prim -> [P a] -> P a

A pattern that matches a primitive expression.

Smart constructors

lam :: Eq a => P a -> Core a -> Core a

smart lam constructor

let_ :: Eq a => [(a, Core a)] -> Core a -> Core a

smart let constructor

alt :: Eq a => P a -> Core a -> Alt () Core a

smart alt constructor

Common built-in terms

cons :: Cons Reviewed Identity s s a a => a -> s -> s

cons an element onto a container.

nil :: Core a

The built-in '[]' constructor for a list.

just :: Core a -> Core a

The built-in Just constructor for Maybe.

nothing :: Core a

The built-in Nothing constructor for Maybe.