Ermine.Syntax.Core

Description

Core Terms

data Branch a

Constructors

Labeled

Fields

tag :: !Int

body :: Scope Int Core a

Default

Fields

body :: Scope Int Core a

data Core a

Core values are the output of the compilation process.

They are terms where the dictionary passing has been made explicit and all of the types have been checked and removed.

Constructors

Var a

HardCore !HardCore

Data !Int [Core a]

App !(Core a) !(Core a)

Lam !Int !(Scope Int Core a)

Let [Scope Int Core a] !(Scope Int Core a)

Case !(Core a) [Branch a]

Dict

Fields

supers :: [Core a]

slots :: [Scope Int Core a]

LamDict !(Scope () Core a)

AppDict !(Core a) !(Core a)

data HardCore

Constructors

Super !Int

Slot !Int

Lit !Literal

class Lit a where

Lifting of literal values to core.

Methods

lit :: a -> Core b

lits :: [a] -> Core b

Smart constructors

lam :: Eq a => [a] -> Core a -> Core a

smart lam constructor

let_ :: Eq a => [(a, Core a)] -> Core a -> Core a

smart let constructor

Common built-in terms

cons :: Cons Reviewed Identity s s a a => a -> s -> s

cons an element onto a container.

nil :: Core a

The built-in '[]' constructor for a list.

just :: Core a -> Core a

The built-in Just constructor for Maybe.

nothing :: Core a

The built-in Nothing constructor for Maybe.