Ermine.Syntax.Core

Description

Core Terms

data Core t a

Core values are the output of the compilation process.

They are terms where the dictionary passing has been made explicit and all of the types have been checked and removed.

Constructors

Var a

HardCore !HardCore

Data [t] !Word64 !Global [Core t a]

Prim [t] t !Global [Core t a]

App !t !(Core t a) !(Core t a)

Lam [t] !(Scope Word64 (Core t) a)

Let [Scope Word64 (Core t) a] !(Scope Word64 (Core t) a)

Case !(Core t a) (Map Word64 (Match t (Core t) a)) (Maybe (Scope () (Core t) a))

Dict

Fields

supers :: [Core t a]

slots :: [Scope Word64 (Core t) a]

data Match t c a

Constructors

Match

Fields

_matchArgs :: [t]

_matchGlobal :: !Global

_matchBody :: Scope Word64 c a

matchArgs :: Lens (Match t c a) (Match t' c a) [t] [t']

matchGlobal :: Lens' (Match t c a) Global

matchBody :: Lens (Match t c a) (Match t d b) (Scope Word64 c a) (Scope Word64 d b)

triverseMatch :: Applicative f => (t -> f t') -> (forall x y. (x -> f y) -> c x -> f (c' y)) -> (a -> f a') -> Match t c a -> f (Match t' c' a')

class (Variable c, AppHash c, AppDict c, App c, Applicative c, Monad c) => Cored t c | c -> t where

Instances of this are things we can both construct from a Core expression, and perform case analysis upon.

e.g. Core and Scope b Core

Minimal complete definition

core, case_, lambda, letrec

Methods

core :: Core t a -> c a

case_ :: c a -> Map Word64 (Match t c a) -> Maybe (Scope () c a) -> c a

lambda :: [t] -> Scope Word64 c a -> c a

letrec :: [Scope Word64 c a] -> Scope Word64 c a -> c a

hardCore :: HardCore -> c a

data JavaLike

Constructors

Method !Bool !Text !Text [Text]
Java methods: static, class name, method name, arg class names

Constructor !Text [Text]
Java constructors: class name, arg class names

Value !Bool !Text !Text
Java values: static, class name, field name

data Foreign

Constructors

JavaLike !JavaLike

Unknown !Text

data HardCore

HardCore is the subset of Core terms that can be unified with value equality.

Constructors

Super !Word64

Slot !Word64

Lit !Literal

Foreign !Foreign

Error !Text

Id !Id

class AsId c => AsHardCore c where

This class describes things that could be HardCore.

Minimal complete definition

Nothing

Methods

_HardCore :: Prism' c HardCore

_Lit :: Prism' c Literal

_Error :: Prism' c Text

_Super :: Prism' c Word64

_Slot :: Prism' c Word64

_Foreign :: Prism' c Foreign

super :: (AsHardCore (c a), AppDict c) => Word64 -> c a -> c a

ask for the nth the superclass of a given dictionary as a core expression

slot :: (AsHardCore (c a), AppDict c) => Word64 -> c a -> c a

ask for the nth slot of a given dictionary as a core expression

Smart constructors

lam :: (Cored t m, Eq a) => t -> [a] -> Core t a -> m a

Smart Lam constructor

let_ :: (Cored t m, Eq a) => [(a, Core t a)] -> Core t a -> m a

Smart Let constructor

dataCon :: Cored t m => [t] -> Word64 -> Global -> m a

Builds an n-ary data constructor

prim :: Cored t m => [t] -> t -> Global -> m a

Builds an n-ary primop