Ermine.Syntax.Global

Description

Documentation

data Global

A Global is a full qualified top level name.

NB: You should construct these with global and only use the constructor for pattern matching.

Constructors

Global !ByteString !Fixity !ByteString !ByteString !ByteString

global :: Fixity -> ByteString -> ByteString -> ByteString -> Global

Construct a Global with a correct digest.

data Assoc

The associativity of an infix identifier

Constructors

L

R

N

data Fixity

The fixity of an identifier

Constructors

Infix !Assoc !Int

Prefix !Int

Postfix !Int

Idfix

Lenses

globalFixity :: Simple Lens Global Fixity

A lens that will read or update the fixity (and compute a new digest)

globalPackage, globalName, globalModule :: Simple Lens Global ByteString

A lenses that will read or update part of the Global and compute a new digest as necessary.

unpackFixity :: Word8 -> Fixity

packFixity :: Fixity -> Word8