Ermine.Syntax.Head

Description

Documentation

data Head

In the instance declaration:

   instance Foo a => Bar Int [a] where
     ...
 

The portion Bar Int [a] is the head of the instance. All instances must have a unique head; type class dispatch does not take the context of the instance into account. Thus, instances are uniquely identified by their heads, and we use them for that purpose among others.

The Head type records all the information necessary for this unique identification, plus a hash for fast equality checks.

The above instance head would be represented as something like:

   Head ... Bar 0 [star] [] [int, list (B 0)]
 

Constructors

Head

Fields

_headDigest :: ByteString

_hash :: !Int

_headClass :: !Global

The class name for which this is an instance

_headBoundKinds :: !Int

number of kind variables brought into scope

_headBoundTypes :: [Kind Int]

kinds of type variables brought into scope

_headKindArgs :: [Kind Int]

_headTypeArgs :: [Type Int Int]

class HasHead t where

Minimal complete definition

head_

Methods

head_ :: Lens' t Head

headClass :: Lens' t Global

headBoundKinds :: Lens' t Int

headBoundTypes :: Lens' t [Kind Int]

headKindArgs :: Lens' t [Kind Int]

headTypeArgs :: Lens' t [Type Int Int]

class AsHead t where

Minimal complete definition

Nothing

Methods

_Head :: Prism' t Head

mkHead :: Global -> Int -> [Kind Int] -> [Kind Int] -> [Type Int Int] -> Head