Ermine.Syntax.Pattern

Description

Documentation

data Pattern t

Patterns used by Term

Constructors

SigP t

WildcardP

AsP (Pattern t)

StrictP (Pattern t)

LazyP (Pattern t)

LitP Literal

ConP Global [Pattern t]

TupP [Pattern t]

_SigP :: forall t. Prism' (Pattern t) t

_WildcardP :: forall t. Prism' (Pattern t) ()

_AsP :: forall t. Prism' (Pattern t) (Pattern t)

_StrictP :: forall t. Prism' (Pattern t) (Pattern t)

_LazyP :: forall t. Prism' (Pattern t) (Pattern t)

_LitP :: forall t. Prism' (Pattern t) Literal

_ConP :: forall t. Prism' (Pattern t) (Global, [Pattern t])

_TupP :: forall t. Prism' (Pattern t) [Pattern t]

_ConP' :: Global -> Prism' (Pattern t) [Pattern t]

data PatHead

Constructors

TupH

Fields

_arity :: Int

ConH

Fields

_arity :: Int

_name :: Global

_TupH :: Prism' PatHead Int

_ConH :: Prism' PatHead (Int, Global)

arity :: Lens' PatHead Int

headName :: Traversal' PatHead Global

traverseHead :: PatHead -> Traversal' (Pattern t) [Pattern t]

data PatPath

Paths into a pattern tree. These will be used as the bound variables for scopes that have patterns in their binder. No effort has been made to enforce statically that pattern paths used in a given scope correspond only to well-formed paths into the particular tree; it will merely have to be an invariant maintained in the compiler.

LeafPP represents a reference to the variable at a node that binds a variable. For instance, LeafPP is a valid path into SigP. It is also valid for AsP p for all p, because AsP binds a variable. It is invalid for ConP, however.

No effort is made to represent paths with no decision points. This is primarily because it is detrimental to the compilation strategy. AsP, StrictP and LazyP can be thought of as annotations that are not full-fledged patterns in themselves. AsP signals that it is valid to make reference to an entire piece of substructure, while StrictP and LazyP inform the pattern compilation process. But it is unnecessary to distinguish between pointing at an AsP---which is always valid---and pointing at the right portion of an AsP---which is invalid unless the pattern contains redundancies.

Constructors

LeafPP
refer to a variable

FieldPP Int PatPath
refer to the n-th subpattern of a constructor

ArgPP Int PatPath
refer to the n-th pattern of many top-level patterns

type PatPaths = Endo PatPath

fieldPP :: Int -> PatPaths

argPP :: Int -> PatPaths

leafPP :: PatPaths -> PatPath

paths :: Pattern t -> [PatPath]

Returns all the paths pointing to variables in the given pattern, in order from left to right.

manyPaths :: [Pattern t] -> [PatPath]

Returns all the paths pointing to variables in a series of patterns. The list is assumed to contain patterns in left-to-right order, and the resulting list will be in the same order.

data Alt t f a

One alternative of a core expression

Constructors

Alt !(Pattern t) !(Scope PatPath f a)

bitraverseAlt :: (Bitraversable k, Applicative f) => (t -> f t') -> (a -> f b) -> Alt t (k t) a -> f (Alt t' (k t') b)

Helper function for traversing both sides of an Alt.

patternHead :: Fold (Pattern t) PatHead

forces :: Pattern t -> Bool

matchesTrivially :: Pattern t -> Bool

serializeAlt3 :: MonadPut m => (t -> m ()) -> (forall a. (a -> m ()) -> f a -> m ()) -> (v -> m ()) -> Alt t f v -> m ()

deserializeAlt3 :: MonadGet m => m t -> (forall a. m a -> m (f a)) -> m v -> m (Alt t f v)