Ermine.Syntax.Pattern.Compiler

Description

Documentation

data Guard c a

Guard representation. Every row of a pattern matrix must have an assocated guard, but if no guard was specified, then it will simply be the trivial guard which allows everything.

Constructors

Trivial

Explicit (Scope PatPath c a)

data PMatrix t c a

Pattern matrices for compilation. The matrix is represented as a list of columns. There is also an extra column representing the guards.

Constructors

PMatrix

Fields

_cols :: [[Pattern t]]

_guards :: [Guard c a]

_bodies :: [Scope PatPath c a]

class HasPMatrix t t c a | t -> t c a where

Methods

pMatrix :: Lens' t (PMatrix t c a)

bodies :: Lens' t [Scope PatPath c a]

cols :: Lens' t [[Pattern t]]

guards :: Lens' t [Guard c a]

newtype PCompEnv

The environment necessary to perform pattern compilation. We need two pieces of information:

1) The signatures (full list of constructors) associated with each particular constructor

2) A mapping from constructors to their Core integer tag

This is accomplished via a map of maps. The outer map should take each global to its associated signature, and signatures are represented as maps from globals to integer tags.

Constructors

PCompEnv

Fields

signatures :: HashMap Global (HashMap Global Word8)

class (Applicative m, Monad m) => MonadPComp m where

Monads that allow us to perform pattern compilation, by providing a PCompEnv.

Methods

askPComp :: m PCompEnv

isSignature :: MonadPComp m => Set PatHead -> m Bool

Determines whether a set of pattern heads constitutes a signature. This is handled specially for tuples and literals, and relies on the monad for data type constructors.

constructorTag :: MonadPComp m => PatHead -> m Word8

Looks up the constructor tag for a pattern head. For tuples this is always 0, but constructors must consult the compilation environment.

data CompileInfo c a

Additional information needed for pattern compilation that does not really belong in the pattern matrix.

The pathMap stores information for resolving previous levels of the pattern to appropriate cores for the current context.

colCores contains core terms that refer to each of the columns of the matrix in the current context.

colPaths contains the paths into the original pattern that correspond to each of the current columns.

Constructors

CInfo

Fields

_pathMap :: HashMap PatPath (c a)

_colCores :: [c a]

_colPaths :: [PatPaths]

class HasCompileInfo t c a | t -> c a where

Methods

compileInfo :: Lens' t (CompileInfo c a)

colCores :: Lens' t [c a]

colPaths :: Lens' t [PatPaths]

pathMap :: Lens' t (HashMap PatPath (c a))

defaultOn :: Applicative c => Int -> PMatrix t c a -> PMatrix t c (Var () (c a))

Computes the matrix that should be used recursively when defaulting on the specified column.

splitOn :: Applicative c => Int -> PatHead -> PMatrix t c a -> PMatrix t c (Var Word8 (c a))

Computes the matrix that should be used recursively when defaulting on the specified column, with the given pattern head.

compile :: (MonadPComp m, Cored c) => CompileInfo c a -> PMatrix t c a -> m (c a)

Compiles a pattern matrix together with a corresponding set of core branches to a final Core value, which will be the decision tree version of the pattern matrix.

compileLambda :: (MonadPComp m, Cored c) => [Pattern t] -> Scope PatPath c a -> m (Scope Word8 c a)