Ermine.Syntax.Term

Description

This module provides the AST for Terms

Terms

data Term t a

Terms in the Ermine language.

Constructors

Var a

App !(Term t a) !(Term t a)

HardTerm !HardTerm

Sig !(Term t a) t

Lam [Pattern t] !(Scope PatPath (Term t) a)

Case !(Term t a) [Alt t (Term t) a]

Let [Binding t a] !(Scope Int (Term t) a)

Loc !Rendering !(Term t a)
informational link to the location the term came from

Remember !Int !(Term t a)
Used to provide hole support.

bindTerm :: (t -> t') -> (a -> Term t' b) -> Term t a -> Term t' b

Perform simultaneous substitution on terms and type annotations.

Hard Terms

data HardTerm

Simple terms that can be compared with structural equality.

Constructors

Lit Literal

DataCon !Global (Type Void Void)

Tuple !Word8

Hole
A placeholder that can take any type. Easy to Remember.

class Terminal t where

This class provides a prism to match against or inject a HardTerm.

Methods

hardTerm :: Prism' t HardTerm

litTerm :: Literal -> t

hole :: t

Bindings

data BodyBound

Bound variables in a declaration are rather complicated. One can refer to any of the following: 1. Definitions in the same declaration sequence 2. Variables bound in a pattern 3. Definitions in a where clause the DeclBound type captures these three cases in the respective constructors.

Constructors

BodyDecl Word32

BodyPat PatPath

BodyWhere Word32

data WhereBound

Constructors

WhereDecl Word32

WherePat PatPath

class AsDecl t where

Methods

_Decl :: Prism' t Word32

data Binding t a

A Binding provides its source location as a rendering, knowledge of if it is explicit or implicitly bound and a list of right hand side bindings.

Constructors

Binding

Fields

_bindingLoc :: !Rendering

_bindingType :: !(BindingType t)

_bindingBodies :: [Body t a]

class HasBinding t t a | t -> t a where

Methods

binding :: Lens' t (Binding t a)

bindingBodies :: Lens' t [Body t a]

bindingLoc :: Lens' t Rendering

bindingType :: Lens' t (BindingType t)

data BindingType t

Indicate if a definition is explicitly bound with a type annotation or implicitly bound without.

Constructors

Explicit t

Implicit

_Implicit :: forall t. Prism' (BindingType t) ()

_Explicit :: forall t t. Prism (BindingType t) (BindingType t) t t

data Body t a

A body is the right hand side of a definition. This isn't a term because it has to perform simultaneous matches on multiple patterns with backtracking. Each Body contains a list of where clause bindings to which the body and guards can refer.

Constructors

Body [Pattern t] (Guarded (Scope BodyBound (Term t) a)) [Binding t (Var WhereBound a)]

bodyDecls :: Traversal' (Body t a) Word32