Ermine.Syntax.Type

Description

Types

data Type k a

Ermine types, parameterized by their free kind variables and free type variables

For the purposes of type checking and general ease of use, we maintain several invariants privileging otherwise isomorphic types. Those invariants are as follows:

There is never a forall to the right of an arrow; the forall is lifted out: T -> forall a. U ==> forall a. T -> U

All variables bound by a quantifier are used: forall a b. a ==> forall a. a

Exists must quantify at least one type or kind variable. Exists 0 [] should be stripped out of all types: exists . C ==> C

Forall must either quantify one variable or contain a non-trivial constraint: forall . T ==> T

Quantifiers must not directly contain another quantifier of the same type; they should be coalesced: forall a. forall b. T ==> forall a b. T

Ands of constraints should not be nested, the lists should be coalesced: ((C, D), E) ==> (C, D, E)

Existentials should be floated out to the top of constraints: (exists a. C, exists b. D) ==> exists a b. (C, D)

Forall should only quantify variables that occur in the non-constraint portion of the body. Other variables should be pushed into existential quantifiers in the constraints: forall a b. C b => a ==> forall a. (exists b. C b) => a

The order of the variables in a forall should correspond to that of their use in the rest of the type. For kinds, this includes their appearance in the annotations of the type variables in the same quantifier:

      forall {l k} (b : l) (a : k) (g : l -> *) (f : k -> *). f a -> g b
        ==>
      forall {k l} (f : k -> *) (a : k) (g : l -> *) (b : l). f a -> g b

Constructors

Var a

App !(Type k a) !(Type k a)

HardType !HardType

Forall [Hint] [Hinted (Scope Int Kind k)] (Scope Int (TK k) a) !(Scope Int (TK k) a)

Loc !Rendering !(Type k a)

Exists [Hint] [Hinted (Scope Int Kind k)] (Scope Int (TK k) a)

And [Type k a]

forall :: (Ord k, Ord t) => (k -> Hint) -> (t -> Hint) -> [k] -> [(t, Kind k)] -> Type k t -> Type k t -> Type k t

A smart constructor for forall. Given a list of kinds, a list of type variables with their kinds, a list of constraints, and a body, abstracts over the kind and type variables in the constraints and the body in the canonical order determined by the body.

exists :: (Ord k, Ord t) => (k -> Hint) -> (t -> Hint) -> [k] -> [(t, Kind k)] -> Type k t -> Type k t

Abstracts over the specified variables using existential quantification. The input type is assumed to meet some invariants (see mergeConstraints), and those invariants are maintained if so.

mergeConstraints :: Type k t -> Type k t -> Type k t

Takes two constraint types and builds their intersection. The arguments are assumed to follow the invariants that: 1) All existentials are outer-most. 2) There are no exists of exists 3) There are no Ands of Ands 4) All quantified variables are actually used in the bodies No attempt is made to simplify the constraints, but the above invariants are maintained.

allConstraints :: [Type k t] -> Type k t

As mergeConstraints, but 0 or more

(~~>) :: Type k t -> Type k t -> Type k t

A smart constructor for function types that moves foralls from the right of an arrow. The right-hand type is assumed to follow the proper invariants for a forall if it is one.

isTrivialConstraint :: Type k a -> Bool

Determines whether the type in question is a trivial constraint, which may be dropped from the type. The simplest example is 'And []', but the function works for non-normalized constraints that are equivalent.

isRowConstraint :: Type k a -> Bool

Determines whether the given type is a row constraint.

As of now, we have no row constraints.

type FieldName = Text

A placeholder for a more complicated database fieldname.

Hard Types

data HardType

A Type that can be compared with mere structural equality.

Constructors

Tuple !Int

Arrow

Con !Global !(Schema Void)

ConcreteRho [FieldName]

class Typical t where

Smart constructors that allows us to pun various HardType constructor names for Type.

Methods

hardType :: Prism' t HardType

tuple :: Int -> t

arrow :: t

con :: Global -> Schema Void -> t

concreteRho :: [FieldName] -> t

Binding

type TK k = Type (Var Int k)

TK is a handy alias for dealing with type scopes that bind kind variables. It's a dumber version than the Bound Scope, as we are unsure that the extra nesting pays off relative to the extra effort.

abstractKinds :: (k -> Maybe Int) -> Type k a -> TK k a

Bind some of the kinds referenced by a Type. N.B. This doesn't do any checking to assure that the integers given are in any kind of canonical order, so this should not be used unless that doesn't matter.

instantiateKinds :: (Int -> Kind k) -> TK k a -> Type k a

Instantiate the kinds bound by a TK obtaining a Type.

instantiateKindVars :: [k] -> TK k a -> Type k a

Instantiate kinds using a list of variables.

bindType :: (k -> Kind k') -> (a -> Type k' b) -> Type k a -> Type k' b

Perform simultaneous substitution on kinds and types in a Type.

bindTK :: (k -> Kind k') -> TK k a -> TK k' a

Substitute kind variables in a TK, leaving the bound kinds untouched.

memoverse :: (Bitraversable t, Applicative f, Monad f, Ord k, Ord a) => (k -> f (Bool, l)) -> (a -> f (Bool, b)) -> t k a -> f (t l b)

A version of bitraverse that allows the functions used to specify whether their results at each value should be remembered, so that the action is not performed additional times at that value.

prepare :: forall t f k a l b. (Bitraversable t, Applicative f, Monad f, Ord k, Ord a) => f l -> (k -> f l) -> (a -> f b) -> t (Maybe k) a -> f (t l b)

A function for preparing the sort of type that comes out of parsing for the inference process. Actions for kind and type variables of types k and t will only be performed once for each unique value, then memoized and the result re-used. The default action for Nothing kinds will be used at every occurrence, however.

abstractAll :: (Ord k, Ord a) => (k -> Hint) -> (a -> Hint) -> Type (Maybe k) a -> (Scope Int (TK ()) b, ([Hint], [Hint]))

Abstract all the unique variables out of a type with ordered type variables and ordered, possibly unknown kind variables. This yields a scope with possibly unknown kind variables and verifiably no type variables. Also returned are the number of unique kind and type variables

Type Variables

class HasTypeVars s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide a Traversal of type variables that can be used to extract them or substitute them for other type variables.

Methods

typeVars :: Traversal s t a b

Type Annotations

data Annot k a

A type annotation

Constructors

Annot [Kind k] !(Scope Int (Type k) a)

annot :: Type k a -> Annot k a

Serialization

serializeTK :: MonadPut m => (k -> m ()) -> (t -> m ()) -> TK k t -> m ()

deserializeTK :: MonadGet m => m k -> m t -> m (TK k t)