Ermine.Unification.Type

Description

This module tells us how to unify types. We use a simplified form of HMF-style unification to deal with unification of foralls.

Documentation

unifyType :: (MonadWriter Any m, MonadMeta s m) => TypeM s -> TypeM s -> m (TypeM s)

Unify two types, with access to a visited set, logging via MonadWriter whether or not the answer differs from the first type argument.

This returns the result of unification with any modifications expanded, as we calculated it in passing

zonkKindsAndTypes :: (MonadMeta s m, MonadWriter Any m) => TypeM s -> m (TypeM s)

zonkKindsAndTypesWith :: (MonadMeta s m, MonadWriter Any m) => TypeM s -> (MetaK s -> m ()) -> (MetaT s -> m ()) -> m (TypeM s)

checkSkolemEscapes :: MonadMeta s m => Maybe Depth -> LensLike' m ts (TypeM s) -> [MetaT s] -> ts -> m ts

Check for escaped Skolem variables in any container full of types.

Dies due to an escaped Skolem or returns the container with all of its types fully zonked.